+++ src/tool/cpp/basewinrt/base.h
@@ -29,7 +29,6 @@
 #include "base_foundation.h"
 #include "base_chrono.h"
 #include "base_security.h"
-#include "base_await.h"
 #include "base_std_hash.h"
 #include "base_std_fire_and_forget.h"
 #include "base_reflect.h"
